src/test/java/com/WiltonZhan/leetcode/l67AddBinary/SolutionTest.java
package com.WiltonZhan.leetcode.l67AddBinary;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class SolutionTest {
    private final Solution solution = new Solution();

    @Test
    void addBinary() {
        assertEquals("111", solution.addBinary("10100000100100110110010000010101111011011001101110111111111101000000101111001110001111100001101",
                "110101001011101110001111100110001010100001101011101010000011011011001011101111001100000011011110011"));
        assertEquals("110100", solution.addBinary("10", "110010"));
        assertEquals("100", solution.addBinary("11", "1"));
        assertEquals("10101", solution.addBinary("1010", "1011"));
    }
}